# Conflicts:
#   backend/src/main/java/com/ssafy/api/service/ChallengeService.java

+++ backend/src/main/java/com/ssafy/api/service/UserService.java
@@ -4,6 +4,7 @@ import com.ssafy.api.dto.UserResDto;
 import com.ssafy.api.request.FollowReqDto;
 import com.ssafy.api.dto.FriendDto;
 import com.ssafy.api.dto.SingleMessageReqDto;
+import com.ssafy.api.response.UserUsageResDto;
 import com.ssafy.common.exception.CustomException;
 import com.ssafy.common.exception.ErrorCode;
 import com.ssafy.db.entity.FriendsList;
@@ -47,6 +48,33 @@ public class UserService {
 				.userProfile(user.getUserProfile())
 				.build();
 		return userResDto;
+	}
+
+	/*
+	 * description : 사용자 기프티콘 등록/사용 조회 메소드
+	 * @param userId : userId
+	 * @return UserUsageResDto :UserUsageResDto 객체
+	 * */
+	public UserUsageResDto getUsage(Long userId){
+		User user = userRepository.findByUserId(userId).orElseThrow(() -> new CustomException(ErrorCode.INVALID_USER));
+		UserUsageResDto userUsageResDto = UserUsageResDto.builder()
+				.registCount(user.getGifticonsList().size())
+				.usedCount(user.getUsedCount())
+				.build();
+		return userUsageResDto;
+	}
+
+	/*
+	 * description : 사용 기프티콘 카운트 업 메소드
+	 * @param userId : userId
+	 * @return boolean : 성공 여부
+	 * */
+	@Transactional
+	public boolean countUpUsed(Long userId){
+		User user = userRepository.findByUserId(userId).orElseThrow(() -> new CustomException(ErrorCode.INVALID_USER));
+		user.countUpUsed();
+		userRepository.save(user);
+		return true;
 	}
 
 	/*
@@ -96,7 +124,8 @@ public class UserService {
 			throw new CustomException(ErrorCode.INVALID_USER);
 		}
 		// 이미 팔로우 했는지 확인
-		if (friendsListRepository.existsByFollowerIdAndFollowingId(followReqDto.getFollowerUserId(), followReqDto.getFollowingUserId())) {
+		if (friendsListRepository.existsByFollowerIdAndFollowingId(followReqDto.getFollowerUserId(), followReqDto.getFollowingUserId())
+		&&!friendsListRepository.findByFollowerIdAndFollowingId(followReqDto.getFollowerUserId(), followReqDto.getFollowingUserId()).get().isFriend()) {
 			FriendsList friendsList = friendsListRepository.findByFollowerIdAndFollowingId(followReqDto.getFollowerUserId(), followReqDto.getFollowingUserId()).get();
 			// 친구 상태 업데이트
 			friendsList.acceptRequest(true);
@@ -134,7 +163,8 @@ public class UserService {
 			throw new CustomException(ErrorCode.INVALID_USER);
 		}
 		// 이미 팔로우 했는지 확인
-		if (friendsListRepository.existsByFollowerIdAndFollowingId(followReqDto.getFollowerUserId(), followReqDto.getFollowingUserId())) {
+		if (friendsListRepository.existsByFollowerIdAndFollowingId(followReqDto.getFollowerUserId(), followReqDto.getFollowingUserId())
+				&&!friendsListRepository.findByFollowerIdAndFollowingId(followReqDto.getFollowerUserId(), followReqDto.getFollowingUserId()).get().isFriend()) {
 			FriendsList friendsList = friendsListRepository.findByFollowerIdAndFollowingId(followReqDto.getFollowerUserId(), followReqDto.getFollowingUserId()).get();
 			friendsListRepository.delete(friendsList);
 			return true;
